Update and question
if the int and float names from the particles are the same names as the detectorhit, should I change the name of the particle int and float ? or leave it the same for now?
ex. x, y, z

+++ src/ScintillatorHit.java
@@ -15,13 +15,13 @@ public class ScintillatorHit extends DetectorHit {
 
 
 
-    public float getlayer() {
+    public int getlayer() {
         return layer;
     } 
     public void setlayer(int layer) {
         this.layer = layer;
     }
-        public float getcomponent() {
+        public int getcomponent() {
         return component;
     } 
     public void setcomponent(int component) {
